pyproject.toml
[project]
name = "cloudlogginginvestigator"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "google-cloud-logging>=3.12.1",
    "pandas>=2.3.3",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "tabulate>=0.9.0",
]
